+++ tools/pylint-extensions/azure-pylint-guidelines-checker/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
-## 0.5.3 (2025-03-19)
+## 0.5.3 (Unreleased)
 - Updating documentation
+- Renaming `do-not-log-exception` to `do-not-log-exception-if-not-debug`
+- Update `do-not-log-raised-errors`
 
 ## 0.5.2 (2025-02-19)